SuccessStories: Refactored Fragments under its own package
