+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/ApplicationFragment.java
@@ -1,4 +1,4 @@
-package com.example.jobs4smcyouth;
+package com.example.jobs4smcyouth.Fragments;
 
 
 import android.os.Bundle;
@@ -8,6 +8,10 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.jobs4smcyouth.ApplicationFragmentRVAdapter;
+import com.example.jobs4smcyouth.R;
+
 import java.util.ArrayList;
 import java.util.List;
