+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/JobPostFragment.java
@@ -1,4 +1,4 @@
-package com.example.jobs4smcyouth;
+package com.example.jobs4smcyouth.Fragments;
 
 
 import android.os.Bundle;
@@ -12,6 +12,7 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.example.jobs4smcyouth.Models.JobListing;
+import com.example.jobs4smcyouth.R;
 import com.firebase.client.DataSnapshot;
 import com.firebase.client.Firebase;
 import com.firebase.client.FirebaseError;
