+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/RequirementFragment.java
@@ -1,4 +1,4 @@
-package com.example.jobs4smcyouth;
+package com.example.jobs4smcyouth.Fragments;
 
 
 import android.os.Bundle;
@@ -9,6 +9,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+
+import com.example.jobs4smcyouth.LaborLawsWebviewFragment;
+import com.example.jobs4smcyouth.R;
 
 
 /**
