+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/TipFragment.java
@@ -1,4 +1,4 @@
-package com.example.jobs4smcyouth;
+package com.example.jobs4smcyouth.Fragments;
 
 
 import android.os.Bundle;
@@ -7,14 +7,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.jobs4smcyouth.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class ScholarshipFragment extends Fragment {
+public class TipFragment extends Fragment {
 
 
-    public ScholarshipFragment() {
+    public TipFragment() {
         // Required empty public constructor
     }
 
@@ -23,7 +25,7 @@ public class ScholarshipFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_scholarship, container, false);
+        return inflater.inflate(R.layout.fragment_tip, container, false);
     }
 
 }
