+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/TransportWebViewFragment.java
@@ -1,4 +1,4 @@
-package com.example.jobs4smcyouth;
+package com.example.jobs4smcyouth.Fragments;
 
 
 import android.graphics.Bitmap;
@@ -10,6 +10,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+
+import com.example.jobs4smcyouth.R;
 
 
 /**
