+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/Fragments/TransportationFragment.java
@@ -1,4 +1,4 @@
-package com.example.jobs4smcyouth;
+package com.example.jobs4smcyouth.Fragments;
 
 
 import android.os.Bundle;
@@ -8,8 +8,8 @@ import android.support.v4.app.FragmentTransaction;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.webkit.WebView;
-import android.widget.FrameLayout;
+
+import com.example.jobs4smcyouth.R;
 
 
 /**
