+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_application.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context="com.example.jobs4smcyouth.ApplicationFragment">
+    tools:context=".Fragments.ApplicationFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
