+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_interview.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context="com.example.jobs4smcyouth.InterviewFragment">
+    tools:context=".Fragments.InterviewFragment">
     <LinearLayout
         android:orientation="vertical"
         android:layout_width="match_parent"
