+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_job_post.xml
@@ -3,7 +3,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.jobs4smcyouth.JobPostFragment">
+    tools:context=".Fragments.JobPostFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <android.support.v7.widget.RecyclerView
