+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_requirement.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context="com.example.jobs4smcyouth.RequirementFragment">
+    tools:context=".Fragments.RequirementFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
