+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_resume.xml
@@ -3,7 +3,7 @@
     xmlns:card_view="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.jobs4smcyouth.ResumeFragment">
+    tools:context=".Fragments.ResumeFragment">
 
     <LinearLayout
         android:layout_marginTop="7dp"
