+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_success_story.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context="com.example.jobs4smcyouth.SuccessStoryFragment">
+    tools:context=".Fragments.SuccessStoryFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/success_stories_rv_id"
