+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_tip.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context="com.example.jobs4smcyouth.TipFragment">
+    tools:context=".Fragments.TipFragment">
 
 <LinearLayout
     android:layout_width="match_parent"
