+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_transportation.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context="com.example.jobs4smcyouth.TransportationFragment">
+    tools:context=".Fragments.TransportationFragment">
 
     <android.support.v7.widget.CardView
         android:layout_width="match_parent"
